refactor: Add command util to open external editor and be used by anothers commands
