+++ src/base/commands/commands/edit_mode.rs
@@ -81,6 +81,21 @@ impl Commands {
         Arc::new(Box::new(S {}))
     }
 
+    pub fn edit_mode_delete_all() -> Command {
+        struct S;
+
+        impl CommandTrait for S {
+            fn execute(&self, app: &mut App) -> Result<(), String> {
+                let input_buffer = app.get_input_buffer_mut();
+                input_buffer.go_to_start();
+                input_buffer.delete_till_end();
+                Ok(())
+            }
+        }
+
+        Arc::new(Box::new(S {}))
+    }
+
     pub fn edit_mode_go_to_prev_char() -> Command {
         struct S;
