+++ src/base/commands/commands/mod.rs
@@ -8,3 +8,4 @@ pub mod submit;
 pub mod tabs;
 pub mod ui;
 pub mod utils;
+pub mod external_editor;
