+++ tests/mod.rs
@@ -1,3 +1,4 @@
 pub mod mocks;
 pub mod integration;
 pub mod e2e;
+pub mod utils;
